Changes for MDG
