+++ src/main/java/net/neoforged/neoform/runtime/engine/NeoFormEngine.java
@@ -6,6 +6,7 @@ import net.neoforged.neoform.runtime.actions.DownloadFromVersionManifestAction;
 import net.neoforged.neoform.runtime.actions.DownloadLauncherManifestAction;
 import net.neoforged.neoform.runtime.actions.DownloadVersionManifestAction;
 import net.neoforged.neoform.runtime.actions.ExternalJavaToolAction;
+import net.neoforged.neoform.runtime.actions.GenerateMCPSrgFilesAction;
 import net.neoforged.neoform.runtime.actions.InjectFromZipFileSource;
 import net.neoforged.neoform.runtime.actions.InjectZipContentAction;
 import net.neoforged.neoform.runtime.actions.MergeWithSourcesAction;
@@ -249,6 +250,21 @@ public class NeoFormEngine implements AutoCloseable {
                     )
             ));
 
+            // We also expose a few results for mappings in different formats
+            var createMappings = graph.nodeBuilder("createMappings");
+            if (hasMojmapSteps) {
+                createMappings.inputFromNodeOutput("officialToObf", "downloadClientMappings", "output");
+                createMappings.inputFromNodeOutput("obfToSrg", "mergeMappings", "output");
+                var action = new CreateLegacyMappingsAction();
+                createMappings.action(action);
+            } else {
+                createMappings.action(new GenerateMCPSrgFilesAction(mcpMappingsData));
+            }
+            graph.setResult(ResultIds.NAMED_TO_INTERMEDIARY_MAPPING, createMappings.output("officialToSrg", NodeOutputType.TSRG, "A mapping file that maps user-facing (Mojang, MCP) names to intermediary (SRG)"));
+            graph.setResult(ResultIds.INTERMEDIARY_TO_NAMED_MAPPING, createMappings.output("srgToOfficial", NodeOutputType.SRG, "A mapping file that maps intermediary (SRG) names to user-facing (Mojang, MCP) names"));
+            graph.setResult(ResultIds.CSV_MAPPING, createMappings.output("csvMappings", NodeOutputType.ZIP, "A zip containing csv files with SRG to official mappings"));
+            createMappings.build();
+
             // If intermediary is in use, the game jar has to be remapped to developer-facing names to be usable
             {
                 var builder = graph.nodeBuilder("remapSrgClassesToOfficial");
@@ -259,24 +275,13 @@ public class NeoFormEngine implements AutoCloseable {
                     builder.input("officialMappings", graph.getRequiredOutput("downloadClientMappings", "output").asInput());
                     builder.action(new RemapSrgClassesAction());
                 } else {
+                    // We can provide the SRG->MCP mappings generated by the earlier task
+                    builder.input("srgToMcpMappings", graph.getResult(ResultIds.INTERMEDIARY_TO_NAMED_MAPPING).asInput());
                     builder.action(new RemapSrgClassesToMcpAction(this.mcpMappingsData));
                 }
                 builder.build();
 
                 graph.setResult(ResultIds.GAME_JAR_NO_RECOMP, officialOutput);
-            }
-
-            if (!processGeneration.classesUseMCPNames()) {
-                // We also expose a few results for mappings in different formats
-                var createMappings = graph.nodeBuilder("createMappings");
-                createMappings.inputFromNodeOutput("officialToObf", "downloadClientMappings", "output");
-                createMappings.inputFromNodeOutput("obfToSrg", "mergeMappings", "output");
-                var action = new CreateLegacyMappingsAction();
-                createMappings.action(action);
-                graph.setResult(ResultIds.NAMED_TO_INTERMEDIARY_MAPPING, createMappings.output("officialToSrg", NodeOutputType.TSRG, "A mapping file that maps user-facing (Mojang, MCP) names to intermediary (SRG)"));
-                graph.setResult(ResultIds.INTERMEDIARY_TO_NAMED_MAPPING, createMappings.output("srgToOfficial", NodeOutputType.SRG, "A mapping file that maps intermediary (SRG) names to user-facing (Mojang, MCP) names"));
-                graph.setResult(ResultIds.CSV_MAPPING, createMappings.output("csvMappings", NodeOutputType.ZIP, "A zip containing csv files with SRG to official mappings"));
-                createMappings.build();
             }
         } else {
             // Without the presence of further patching or renaming, the game jar without recompilation is the deobfuscated vanilla jar
